### 2015-11-26 #22 @AndySouth
rnaturalearth
* 1. What does this package do? (explain in 50 words or less)
rnaturalearth provides :
1. a pre-downloaded subset of Natural Earth vectors commonly used in world mapping.
2. functions to download other Natural Earth vectors.
3. an open, reproducible workflow from www.naturalearthdata.com enabling updating as new versions become available.
4. flexibility to get maps classified by countries, sovereign states and map-units.

* 2. Paste the full DESCRIPTION file inside a code block (bounded by ``` on either end).
```
Package: rnaturalearth
Title: World Vector Map Data from Natural Earth
Version: 0.0.0.9000
Authors@R: person("Andy", "South", , "southandy@gmail.com", role = c("aut", "cre"))
Description: Facilitates mapping by making natural earth map data from http://
    www.naturalearthdata.com/ more easily available to R users. Focuses on vector
    data.
License: CC0
LazyDataCompression: xz
    sp (>= 1.0.15),
    rgdal
    httr
```
* 3. URL for the package (the development repository, not a stylized html page)
https://github.com/AndySouth/rnaturalearth
* 4. What data source(s) does it work with (if applicable)?
www.naturalearthdata.com
* 5. Who is the target audience?
Anyone who wants to use Natural Earth vector data to make a map or perform other geographical analyses.
* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?
More limited subsets of Natural Earth data are available in at least the following CRAN packages rworldmap, rworldxtra, choroplethr, maps, oce, tmap. rnaturalearth is different in that it provides a comprehensive reproducible solution for access to Natural Earth vector map data either pre-downloaded or by facilitating download. By separating data access from visualisation this package provides a resource that can be used by other visualisation packages.

* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
* [x] This package does not violate the Terms of Service of any service it interacts with.
* [x] The repository has continuous integration with Travis and/or another service
* [x] The package contains a vignette
* [x] The package contains a reasonably complete readme with devtools install instructions
* [x] The package contains unit tests
* [x] The package only exports functions to the NAMESPACE that are intended for end users
* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
* [no] Are there any package dependencies not on CRAN?
* [yes] Do you intend for this package to go on CRAN?
* [yes] Does the package have a CRAN accepted license?
* [no] Did `devtools::check()` produce any errors or warnings? If so paste them below.
One NOTE
* checking installed package size ... NOTE
    data  25.8Mb

* 9. Please add explanations below for any exceptions to the above:
The package is currently too big because of the fine scale (10m) data. Hadley suggested :
"I think you need to keep the data under five meg. I'd suggest that you put the fine level data a separate package - you can then make that available via a drat repo (perhaps the ROpenSci one?), or just via github."
I don't know anything about drat repos. I'm happy to take advice and modify the package(s) as recommended.
