### 2015-11-26 #22 @sckott
@AndySouth Thanks for this!  Seeking reviewer(s)
