### 2015-11-30 #22 @sckott
@Robinlovelace and @lmullen reviewing - thanks both!
