### 2015-12-02 #22 @AndySouth
Thanks all. I appreciate your efforts.
If you have questions feel free to ask.

On 30 November 2015 at 16:17, Scott Chamberlain <notifications@github.com>
wrote:

> @Robinlovelace <https://github.com/Robinlovelace> and @lmullen
> <https://github.com/lmullen> reviewing - thanks both!
>
