### 2015-12-15 #22 @AndySouth
Thanks @lmullen for a very helpful review and for giving me clear
guidelines to follow. I knew some of these needed doing, now I know how.
I'll wait for feedback from others and then get stuck in.

On 14 December 2015 at 22:31, Lincoln Mullen <notifications@github.com>
wrote:

> First, let me say that this package is very useful. It will be a very
> helpful addition to the mapping packages on CRAN and in rOpenSci. I already
> have plans to use it for my own work and in assignments for students. Well
> done, @AndySouth <https://github.com/AndySouth>.
>
> It takes about 2 minutes to install the dev version from GitHub and the
> installed version is 26 MB. In addition to the fact that CRAN won't accept
> the package until the data size is smaller, the weight of this package also
> hinders you as the developer from doing updates as necessary, and it makes
> it difficult for users to install the package. To solve this it will be
> necessary to break this package into two or more other packages as detailed
> below. For the purposes of this review I've focused mainly on package-level
> issues. Once you've solved the size problem and the package design has
> stabilized, I'll review the code more granularly.
> Structural changes
>
> This package should be at least two packages, and probably three. Here is
> a suggestion for how you should break this up:
>
>    - The main package would be rnaturalearth and it would be on CRAN. It
>    would be the only package that a user would ever install or load directly.
>    It should contain nothing but the code to download NaturalEarth files,
>    functions to access the data from the other packages, and the vignettes and
>    package level documentation. The absence of data from this package means
>    that you can update it with impunity, since the size of the package will be
>    very small. (I've found this also helps speed up development.)
>    - The second package would be rnaturalearthData or the like, and it
>    should also be on CRAN. It would contain nothing but the most essential
>    data: as much as you can fit while keeping the package under the 5MB limit
>    set by CRAN. It will also contain the documentation for the datasets. This
>    package will be imported by rnaturalearth, so users will install it
>    whenever they install rnaturalearth. But since this data will changed
>    infrequently, users will seldom need to update this package.
>    - The third package will be rnaturalearthHiRes or the like. This
>    package would reside solely in the rOpenSci drat repository
>    <http://packages.ropensci.org/>. It would be a suggested package of
>    rnaturalearth, using the AdditionalRepos field in the DESCRIPTION.
>    When users call a function that would use this pre-downloaded data, the
>    rnaturalearth package would check if rnaturalearthHiRes is installed.
>    If it is not installed, then it will ask to install the package from the
>    rOpenSci drat repo. Since this would be a non-CRAN package, you can include
>    as much data as you want. Whether it would ever make sense to break this
>    high resolution package into several different packages would be up to you.
>
> A compromise solution would be to put minimal data into rnaturalearth,
> and have all the other data in a single rOpenSci drat repository package.
> But if you're going to do this, I think you may as well get as well break
> it up into one further package as suggested above.
>
> Now that is a lot of breaking up into multiple packages what will remain,
> as far as users are concerned, a single user-facing package. But this is
> the accepted practice for packages which depend on a lot of data. For
> instance, the openNLP package installs very large packages containing
> trained models for different langauges from Datacube. I've had to do this
> in gender and genderdata, even though the data there is much smaller than
> the geographic data you are working with. And I'm in the process of redoing
> USAboundaries into exactly the same package structure that I'm suggesting
> here. If it is a helpful model, you might take a look at gender
> <https://github.com/ropensci/gender> and genderdata
> <https://github.com/ropensci/genderdata>. Here are the relevant functions
> <https://github.com/ropensci/gender/blob/master/R/install-genderdata-package.R>
> .
>
> When you do this, you should feel free to give the data packages the CC0
> license, but to give the rnaturalearth package which contains only your
> own work whatever open source license you prefer.
>
> @sckott <https://github.com/sckott>, @karthik <https://github.com/karthik>,
> @cboettig <https://github.com/cboettig>: To be able to test these changes
> effectively, @AndySouth <https://github.com/AndySouth> will need to use
> the rOpenSci drat repository. I'd suggest that the package be provisionally
> accepted and moved into rOpenSci GitHub and drat repositories.
> Tooling
>
>    - The README.md should be converted to README.Rmd via
>    devtools::use_readme_rmd(). As a bonus, this will let the README
>    include images easily.
>    - The package should use Appveyor's CI for Windows via
>    devtools::use_appveyor().
>
> DESCRIPTION
>
>    - Specify a version number for rgdal, httr, and knitr.
>
> Tests
>
> All the tests pass. I'll check coverage more carefully once the package
> structure has stabilized. In the meantime, all of the tests which download
> files should have skip_on_cran().
> Vignette
>
> The vignette should not need download anything from Natural Earth. The
> code block at the end with two downloads should be set to eval = FALSE.
> Code style
>
> There are places where the code style can be improved. See the suggestions
> in this chapter of Hadley Wickham's *R Packages* book
> <http://r-pkgs.had.co.nz/r.html> or the style chapter
> <http://adv-r.had.co.nz/Style.html> in his *Advanced R* book.
>
>    - In the README and vignettes, in place of require(rnaturalearth) use
>    library(rnaturalearth).
>    - Spacing is often inconsistent around = using in function definitions
>    and calls. There should always be a single space on both sides of =.
>    For example, in the README vignette("rnaturalearth",
>    package="rnaturalearth") should be vignette("rnaturalearth", package =
>    "rnaturalearth").
>    - Comments at the start of the line should be followed by a space. (I
>    prefer to user sentence capitalization; but that's just a preference.) E.g.,
>
> # This is more legible
>
>
>    - Most important, all code which has been commented out needs to be
>    deleted.
>    -
>
>    The same goes for the data download scripts in data-raw. At the
>    moment, they have a lot investigating the data. Now that you've settled on
>    what to do, you can retain that information in comments but trim
>    unnecessary code.
>    -
>
>    Functions which have limited categorical options should specify all
>    the options in the function signature and then use match.args(). An
>    example of this is get_data() and the type parameter. For example, a
>    function that looks like this
>
> myfunc <- function(type = "a") {
>   if(type == "a") {
>     # ...
>   } else if (type =="b") {
>
>   } else {
>     stop("Error.")
>   }
> }
>
> Should be changed to this:
>
> myfunc <- function(type = c("a", "b")) {
>   type <- match.arg(type)
>   # Logic checking type here.
> }
>
>
>    - get_data() should be renamed to something less generic, like
>    ne_get_data(). Or, can you avoid exporting it?
>
> Human language style
>
>    - In the headings of the vignettes, prefer sentence-style
>    capitalization. E.g. 1. Maps in the package. (I'd suggest the same for
>    comments as well, but that's entirely up to you.)
>    - This line in the vignette should be revised, perhaps into a bulleted
>    list, since a sentence can only contain one colon. contains
>    pre-downloaded vector maps for countries: ne_countries(), states:
>    ne_states() and coastline: ne_coastline()
>
