### 2015-12-17 #22 @AndySouth
Thanks @Robinlovelace <https://github.com/Robinlovelace> for a considered
and useful review.
