### 2015-12-18 #22 @richfitz
Another option (though experimental and could delay CRAN submission).

In another context, I'm developing a framework for using GitHub to distribute evolving data sources without putting them in a package; see [early vignette](http://richfitz.github.io/datastorr/vignettes/datastorr.html), and [beginnings of manuscript](https://github.com/richfitz/data_versioning/blob/master/ms.tex) which include a little about where we're thinking of taking this.

In short, the data don't go in the package, or even in the repository but live in "github releases" and are fetched as they're used.  This decouples the development of the code of the package from the data source which avoids the issues of duplicating the data over and over on CRAN and makes for fast package downloads.
