### 2015-12-22 #22 @lmullen
@richfitz: datastorr looks really great. I'm looking forward to the release.
