### 2016-02-12 #22 @AndySouth
Thanks @sckott <https://github.com/sckott>. I've been meaning to get back
to this.
I have a plan to start by following parts of the suggestions of both
reviewers. I'll try to get something to you in the next few weeks.

On 11 February 2016 at 19:34, Scott Chamberlain <notifications@github.com>
wrote:

> hey @AndySouth <https://github.com/AndySouth> let us know if there's
> anything we can do to help...
>
