### 2016-02-13 f3d5c03
delete datasets and asscociated doc moved to rnaturalearthdata and rnaturalearthhires
