+++ man/sovereignty.Rd
@@ -1,35 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/datasets.r
-\docType{data}
-\name{sovereignty}
-\alias{sovereignty}
-\alias{sovereignty10}
-\alias{sovereignty110}
-\alias{sovereignty50}
-\title{world sovereignty polygons from Natural Earth}
-\format{A \code{SpatialPolygonsDataFrame}}
-\source{
-\url{http//www.naturalearthdata.com/download/110m/cultural/ne_110m_admin_0_sovereignty.zip}
-
-\url{http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_sovereignty.zip}
-
-\url{http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_sovereignty.zip}
-}
-\usage{
-sovereignty110
-
-sovereignty50
-
-sovereignty10
-}
-\description{
-sovereignty polygons at scales 1:110m (small), 1:50m (medium), 1:10m (large).   
-There are fewer sovereign states than countries e.g. the Falkland Islands are included with the United Kingdom.
-}
-\section{Slots}{
-
-\describe{
-\item{\code{data}}{A data frame with attributes}
-}}
-\keyword{datasets}
-
