### 2016-02-13 e1cad1e
change licence to MIT
