+++ DESCRIPTION
@@ -5,7 +5,7 @@ Authors@R: person("Andy", "South", , "southandy@gmail.com", role = c("aut", "cre
 Description: Facilitates mapping by making natural earth map data from http://
     www.naturalearthdata.com/ more easily available to R users. Focuses on vector
     data.
-License: CC0
+License: MIT
 LazyData: true
 LazyDataCompression: xz
 URL: https://github.com/AndySouth/rnaturalearth
