+++ R/get_data.r
@@ -13,6 +13,9 @@
 get_data <- function(scale = 110,
                      type = 'countries') {
 
+  # check for the data package and try to install if not there
+  # I may want to avoid this check for one example dataset in this package (i.e. countries110)
+  check_rnaturalearthdata()
 
   # check on permitted scales, convert names to numeric
   scale <- check_scale(scale)
@@ -21,7 +24,6 @@ get_data <- function(scale = 110,
   #but before I do that I want to sort which datasets are going to be in the package
   #e.g. some type,scale combinations are not available
   #getExportedValue("rnaturalearth", paste0(type,scale))
-  
   
   # choose which map based on type and scale
   # i could use paste to build up varname but this may be safer
