+++ man/check_rnaturalearthdata.Rd
@@ -1,0 +1,13 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/install-naturalearthdata.r
+\name{check_rnaturalearthdata}
+\alias{check_rnaturalearthdata}
+\title{Check whether to install rnaturalearthdata and install if necessary}
+\usage{
+check_rnaturalearthdata()
+}
+\description{
+If the rnaturalearthdata package is not installed, install it from GitHub using
+devtools. If it is not up to date, reinstall it.
+}
+
