+++ man/install_naturalearthdata.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/install-naturalearthdata.r
+\name{install_naturalearthdata}
+\alias{install_naturalearthdata}
+\title{Install the naturalearthdata package after checking with the user}
+\usage{
+install_naturalearthdata()
+}
+\description{
+Install the naturalearthdata package after checking with the user
+}
+
