+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
+export(check_rnaturalearthdata)
+export(install_naturalearthdata)
 export(ne_coastline)
 export(ne_countries)
 export(ne_download)
