### 2016-02-16 f6b0bb1
removed mention of datafiles moved to other packages
