### 2016-02-16 eb0243a
tweaked description
