### 2016-02-16 6a78648
trying to fix data package installation
