+++ R/install-rnaturalearthdata.r
@@ -16,7 +16,7 @@ check_rnaturalearthdata <- function() {
 
 #' Install the naturalearthdata package after checking with the user
 #' @export
-install_naturalearthdata <- function() {
+install_rnaturalearthdata <- function() {
   instructions <- paste(" Please try installing the package for yourself",
                         "using the following command: \n",
                         'devtools::install_github("AndySouth/rnaturalearthdata")' )
