+++ man/install_rnaturalearthdata.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/install-rnaturalearthdata.r
+\name{install_rnaturalearthdata}
+\alias{install_rnaturalearthdata}
+\title{Install the naturalearthdata package after checking with the user}
+\usage{
+install_rnaturalearthdata()
+}
+\description{
+Install the naturalearthdata package after checking with the user
+}
+
