+++ man/check_rnaturalearthdata.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/install-naturalearthdata.r
+% Please edit documentation in R/install-rnaturalearthdata.r
 \name{check_rnaturalearthdata}
 \alias{check_rnaturalearthdata}
 \title{Check whether to install rnaturalearthdata and install if necessary}
