+++ man/install_rnaturalearthhires.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/install-naturalearthdata.r
-\name{install_naturalearthdata}
-\alias{install_naturalearthdata}
+% Please edit documentation in R/install-rnaturalearthhires.r
+\name{install_rnaturalearthhires}
+\alias{install_rnaturalearthhires}
 \title{Install the naturalearthdata package after checking with the user}
 \usage{
-install_naturalearthdata()
+install_rnaturalearthhires()
 }
 \description{
 Install the naturalearthdata package after checking with the user
