+++ NAMESPACE
@@ -1,7 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
 export(check_rnaturalearthdata)
-export(install_naturalearthdata)
+export(check_rnaturalearthhires)
+export(install_rnaturalearthdata)
+export(install_rnaturalearthhires)
 export(ne_coastline)
 export(ne_countries)
 export(ne_download)
