### 2016-02-16 93d4cff
fix data checks
