### 2016-02-16 a1044ff
comment out download within vignette
