+++ vignettes/rnaturalearth.Rmd
@@ -84,12 +84,12 @@ Each [Natural Earth](http://www.naturalearthdata.com/) dataset is characterised 
 ```{r, eval=TRUE, echo=TRUE, message=FALSE}
 
 #lakes
-lakes110 <- ne_download(scale=110, type='lakes', category='physical')
-sp::plot(lakes110, col='blue')
+#lakes110 <- ne_download(scale=110, type='lakes', category='physical')
+#sp::plot(lakes110, col='blue')
 
 #rivers
-rivers110 <- ne_download(scale=110, type='rivers_lake_centerlines', category='physical')
-sp::plot(rivers110, col='blue')
+#rivers110 <- ne_download(scale=110, type='rivers_lake_centerlines', category='physical')
+#sp::plot(rivers110, col='blue')
 
 
 ```
