### 2016-02-16 73fa6f6
add checks on passed filters
