### 2016-02-17 6357a63
add back in raster option
