### 2016-02-17 4d7c89b
added appveyor CI
