### 2016-03-02 a8e7ad7
add to doc about limks between ne_load and download
