### 2016-03-02 d2b8834
add details about available data to man
