+++ man/rnaturalearth.Rd
@@ -4,11 +4,11 @@
 \name{rnaturalearth}
 \alias{rnaturalearth}
 \alias{rnaturalearth-package}
-\title{rnaturalearth : world vector maps from Natural Earth}
+\title{rnaturalearth : world map data from Natural Earth}
 \description{
 Facilitates world mapping by making \href{http://www.naturalearthdata.com/}{Natural Earth} map data more easily available to R users.
 }
 \seealso{
-ne_countries
+\code{\link{ne_countries}} \code{\link{ne_download}}
 }
 
