### 2016-03-02 34d5a61
replace require w library suggested by Lincoln
