+++ vignettes/rnaturalearth.Rmd
@@ -27,8 +27,8 @@ This vignette uses `sp::plot` as a simple, quick way to show how different data 
 
 #### load required packages
 ```{r, eval=TRUE, echo=TRUE, message=FALSE}
-require(rnaturalearth)
-require(sp)
+library(rnaturalearth)
+library(sp)
 ```
 
 
