+++ man/ne_download.Rd
@@ -30,13 +30,35 @@ if \code{destdir} is specified the data can be reloaded in a later R session usi
 with the same arguments.
 }
 \details{
-here is a non exhaustive list of datasets available according to \code{scale} specified by the \code{type} param 
+A non-exhaustive list of datasets available according to \code{scale} specified by the \code{type} param 
   \tabular{lccc}{
-  type="…"	         \tab scale=110	\tab scale=50	\tab scale=10 \cr
+         	                   \tab scale=110	\tab scale=50	\tab scale=10 \cr
   category="Physical" \cr
-  coastline	       \tab y	        \tab y      	\tab y        \cr
-  land     	       \tab y	        \tab y      	\tab y        \cr
-  ocean     	       \tab y	        \tab y      	\tab y        \cr   
+  type=""             \cr
+  coastline	                 \tab y	        \tab y      	\tab y        \cr
+  land     	                 \tab y	        \tab y      	\tab y        \cr
+  ocean     	                 \tab y	        \tab y      	\tab y        \cr
+  rivers_lake_centerlines     \tab y	        \tab y      	\tab y        \cr
+  lakes     	                 \tab y	        \tab y      	\tab y        \cr 
+  glaciated_areas     	       \tab y	        \tab y      	\tab y        \cr
+  antarctic_ice_shelves_polys \tab y	        \tab y      	\tab y        \cr
+  geographic_lines            \tab y	        \tab y      	\tab y        \cr
+  graticules_1     	         \tab y	        \tab y      	\tab y        \cr
+  graticules_30     	         \tab y	        \tab y      	\tab y        \cr
+  wgs84_bounding_box     	   \tab y	        \tab y      	\tab y        \cr
+  playas     	               \tab -	        \tab y      	\tab y        \cr
+  minor_islands      	       \tab -	        \tab -      	\tab y        \cr
+  reefs              	       \tab -	        \tab -      	\tab y        \cr 
+  category="Cultural" \cr
+  type=""             \cr    
+  populated_places        	   \tab y	        \tab y      	\tab y        \cr
+  admin_0_boundary_lines_land \tab y	        \tab y      	\tab y        \cr
+  admin_0_breakaway_disputed_areas \tab -	        \tab y      	\tab y        \cr
+  airports              	     \tab -	        \tab y      	\tab y        \cr
+  ports              	       \tab -	        \tab y      	\tab y        \cr
+  urban_areas              	 \tab -	        \tab y      	\tab y        \cr
+  roads              	       \tab -	        \tab -      	\tab y        \cr   
+  railroads              	   \tab -	        \tab -      	\tab y        \cr       
   }
 }
 \examples{
@@ -65,6 +87,6 @@ spdf_world2 <-    ne_load( scale = 110, type = 'countries' )
  
 }
 \seealso{
-\code{\link{ne_load}}
+\code{\link{ne_load}}, pre-downloaded data are available using \code{\link{ne_countries}}, \code{\link{ne_states}}
 }
 
