+++ man/rnaturalearth.Rd
@@ -9,6 +9,6 @@
 Facilitates world mapping by making \href{http://www.naturalearthdata.com/}{Natural Earth} map data more easily available to R users.
 }
 \seealso{
-\code{\link{ne_countries}} \code{\link{ne_download}}
+\code{\link{ne_countries}} \code{\link{ne_states}} \code{\link{ne_download}}
 }
 
