### 2016-03-02 8c21f66
set download chunk to eval=FALSE
