### 2016-03-02 cb11d38
add smallmediumlarge to scale man options
