+++ R/ne_load.r
@@ -3,7 +3,7 @@
 #' returns loaded data as a spatial object.
 #' 
 #'
-#' @param scale scale of map to return, one of \code{110}, \code{50}, \code{10}
+#' @param scale scale of map to return, one of \code{110}, \code{50}, \code{10} or \code{'small'}, \code{'medium'}, \code{'large'}
 #' @param type type of natural earth file one of 'countries', 'map_units', 'map_subunits', 'sovereignty', 'states'
 #'    OR the portion of any natural earth vector url after the scale and before the . 
 #'    e.g. for "ne_50m_urban_areas.zip" this would be "urban_areas"
