+++ man/ne_load.Rd
@@ -8,7 +8,7 @@ ne_load(scale = 110, type = "countries", category = c("cultural",
   "physical", "raster"), destdir = tempdir(), file_name = NULL)
 }
 \arguments{
-\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10}}
+\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10} or \code{'small'}, \code{'medium'}, \code{'large'}}
 
 \item{type}{type of natural earth file one of 'countries', 'map_units', 'map_subunits', 'sovereignty', 'states'
 OR the portion of any natural earth vector url after the scale and before the . 
