+++ tests/testthat/test-ne_file_name.r
@@ -33,6 +33,8 @@ url_expect_fail <- ne_file_name(scale=110, type='expect_fail', category='cultura
 
 test_that("urls for downloads created by the package exist", {
   
+  testthat::skip_on_cran()
+  
   #tests all of the urls put into the vector above
   #info=x means that test fail messages include the failed urls
   sapply(urls, function(x) expect_true(httr::url_success(x), info=x))
@@ -40,6 +42,8 @@ test_that("urls for downloads created by the package exist", {
   })
 
 test_that("a bogus url does not exist", {
+  
+  testthat::skip_on_cran()
   
   expect_false(httr::url_success(url_expect_fail))
   
