### 2016-03-05 56fb2e7
tidied, removed parts relating to files now stored in packages rnaturalearthdata and rnaturalearthhires
