### 2016-03-09 3181a4f
devtools::use_readme_rmd(), copy text from readme.md to readme.Rmd and knit
