+++ README.Rmd
@@ -1,67 +1,68 @@
+---
+output:
+  md_document:
+    variant: markdown_github
+---
+
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+
+```{r, echo = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "README-"
+)
+```
+
 [![Travis-CI Build Status](https://travis-ci.org/AndySouth/rnaturalearth.svg?branch=master)](https://travis-ci.org/AndySouth/rnaturalearth)
 
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/AndySouth/rnaturalearth?branch=master&svg=true)](https://ci.appveyor.com/project/AndySouth/rnaturalearth)
 
-rnaturalearth
-=============
+# rnaturalearth
 
 An R package to hold and facilitate interaction with [Natural Earth](http://www.naturalearthdata.com/) vector map data.
 
 ### Aims :
+1. access to a pre-downloaded subset of Natural Earth vectors commonly used in world mapping
+1. allow easy subsetting by countries and regions
+1. facilitate download of other Natural Earth vectors
+1. provide a simple, reproducible and sustainable workflow from Natural Earth data to rnaturalearth enabling updating as new versions become available
+1. to clarify differences in world maps classified by countries, sovereign states and map units
+1. to follow Natural Earth naming conventions so that rnaturalearth users can use Natural Earth documentation
 
-1.  access to a pre-downloaded subset of Natural Earth vectors commonly used in world mapping
-2.  allow easy subsetting by countries and regions
-3.  facilitate download of other Natural Earth vectors
-4.  provide a simple, reproducible and sustainable workflow from Natural Earth data to rnaturalearth enabling updating as new versions become available
-5.  to clarify differences in world maps classified by countries, sovereign states and map units
-6.  to follow Natural Earth naming conventions so that rnaturalearth users can use Natural Earth documentation
-
-The [Natural Earth](http://www.naturalearthdata.com/) website structures vector data by scale, category and type. These determine the filenames of downloads. rnaturalearth uses this structure to facilitate download (like an API).
+The [Natural Earth](http://www.naturalearthdata.com/) website structures vector data by scale, category and type. These determine the filenames of downloads. rnaturalearth uses this structure to facilitate download (like an API). 
 
 ### Install rnaturalearth
 
-    devtools::install_github("andysouth/rnaturalearth", build_vignettes=TRUE)
+```
+devtools::install_github("andysouth/rnaturalearth", build_vignettes=TRUE)
+
+```
 
 ### First Usage
-
 Here using `sp::plot` as a simple, quick way to plot maps. Maps could also be made with `ggplot2`, `tmap` or other options.
-
-``` r
+```{r, eval=TRUE}
 library(rnaturalearth)
 library(sp)
 
 #world countries
 sp::plot(ne_countries())
-```
-
-![](README-unnamed-chunk-2-1.png)<!-- -->
-
-``` r
 #uk
 sp::plot(ne_countries(country = 'united kingdom'))
-```
-
-![](README-unnamed-chunk-2-2.png)<!-- -->
-
-``` r
 #states, admin level1 boundaries
 sp::plot(ne_states(country ='spain')) 
+
 ```
 
-![](README-unnamed-chunk-2-3.png)<!-- -->
-
 ### Introductory vignette
-
-``` r
+```{r, eval=FALSE}
 vignette("rnaturalearth", package="rnaturalearth")
 ```
 
 ### To download Natural Earth data not already in the package
-
[…]
