+++ R/ne_load.r
@@ -13,30 +13,28 @@
 
 #' @seealso \code{\link{ne_download}}
 #' @examples
-#' #commented out to stop download in check
-#' #download followed by load from tempdir() works in same R session
-#' #spdf_world <- ne_download( scale = 110, type = 'countries' )
-#' #spdf_world2 <-    ne_load( scale = 110, type = 'countries' )
-#' #download followed by load from specified directory works between R sessions
-#' #spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
-#' #spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )
-#' #for raster
-#' #download & load
-#' #rst <- ne_download(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
-#' #load after having downloaded
-#' #rst <- ne_load(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
-#' #plot
-#' #library(raster)
-#' #plot(rst)
+#' # commented out to stop download in check
+#' # download followed by load from tempdir() works in same R session
+#' # spdf_world <- ne_download( scale = 110, type = 'countries' )
+#' # spdf_world2 <-    ne_load( scale = 110, type = 'countries' )
+#' # download followed by load from specified directory works between R sessions
+#' # spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
+#' # spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )
+#' # for raster
+#' # download & load
+#' # rst <- ne_download(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' # load after having downloaded
+#' # rst <- ne_load(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' # plot
+#' # library(raster)
+#' # plot(rst)
 
 #' @return A \code{Spatial} object depending on the data (points, lines, polygons or raster).
 #' @export
 
 ne_load <- function(scale = 110,
                     type = 'countries',
-                    #category = c('cultural', 'physical'),
                     category = c('cultural', 'physical', 'raster'),
-                    #destdir = NULL,
                     destdir = tempdir(),
                     file_name = NULL
 ) 
@@ -67,7 +65,7 @@ ne_load <- function(scale = 110,
   } else #for shapefiles
   {
     
-    #add '.shp' for the exists test (it's not needed by readOGR)
+    # add '.shp' for the exists test (it's not needed by readOGR)
     if (!file.exists( file.path(destdir, paste0(file_name,'.shp'))))
       stop(error_msg)
     
