+++ man/check_scale.Rd
@@ -16,9 +16,9 @@ integer scale of map
 check name or numeric scale representations, return numeric one
 }
 \examples{
-#commented out because not exported
-#check_scale(110)
-#check_scale("small")
+# commented out because not exported
+# check_scale(110)
+# check_scale("small")
 
 }
 
