### 2016-03-09 456e6b2
add dontrun to examples
