+++ R/ne_load.r
@@ -13,22 +13,27 @@
 
 #' @seealso \code{\link{ne_download}}
 #' @examples
-#' # commented out to stop download in check
+#' \dontrun{
 #' # download followed by load from tempdir() works in same R session
-#' # spdf_world <- ne_download( scale = 110, type = 'countries' )
-#' # spdf_world2 <-    ne_load( scale = 110, type = 'countries' )
+#' spdf_world <- ne_download( scale = 110, type = 'countries' )
+#' spdf_world2 <-    ne_load( scale = 110, type = 'countries' )
+#' 
 #' # download followed by load from specified directory works between R sessions
-#' # spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
-#' # spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )
+#' spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
+#' spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )
+#' 
 #' # for raster
 #' # download & load
-#' # rst <- ne_download(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' rst <- ne_download(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' 
 #' # load after having downloaded
-#' # rst <- ne_load(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' rst <- ne_load(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' 
 #' # plot
-#' # library(raster)
-#' # plot(rst)
-
+#' library(raster)
+#' plot(rst)
+#' } # end dontrun
+#' 
 #' @return A \code{Spatial} object depending on the data (points, lines, polygons or raster).
 #' @export
 
