+++ R/ne_download.r
@@ -42,6 +42,7 @@
 #'   }
 #' @seealso \code{\link{ne_load}}, pre-downloaded data are available using \code{\link{ne_countries}}, \code{\link{ne_states}}
 #' @examples
+#' \dontrun{
 #' spdf_world <- ne_download( scale = 110, type = 'countries' )
 #' 
 #' if (require(sp)) {
@@ -53,18 +54,20 @@
 #' spdf_world2 <-    ne_load( scale = 110, type = 'countries' )
 #' 
 #' # download followed by load from specified directory will work across sessions
-#' # spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
-#' # spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )
+#' spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
+#' spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )
 #'
 #' # for raster
 #' # download & load
-#' # rst <- ne_download(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' rst <- ne_download(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' 
 #' # load after having downloaded
-#' # rst <- ne_load(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' rst <- ne_load(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' 
 #' # plot
-#' # library(raster)
-#' # plot(rst)
-#'  
+#' library(raster)
+#' plot(rst)
+#' } # end dontrun
 #' @return A \code{Spatial} object depending on the data (points, lines, polygons or raster), 
 #'    unless load=FALSE in which case it returns the name of the downloaded shapefile (without extension).
 #' @export
