### 2016-03-09 7072ae9
add 10 & large as scale options
