+++ man/ne_countries.Rd
@@ -1,6 +1,7 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/ne_countries.r
 \name{ne_countries}
+\alias{ne_admin0}
 \alias{ne_countries}
 \title{Get natural earth world country polygons}
 \usage{
@@ -8,7 +9,7 @@ ne_countries(scale = 110, type = "countries", continent = NULL,
   country = NULL, geounit = NULL, sovereignty = NULL)
 }
 \arguments{
-\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{'small'}, \code{'medium'}}
+\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10} or \code{'small'}, \code{'medium'}, \code{'large'}}
 
 \item{type}{country type, one of 'countries', 'map_units', 'sovereignty'}
 
