+++ man/ne_states.Rd
@@ -1,6 +1,7 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/ne_states.r
 \name{ne_states}
+\alias{ne_admin1}
 \alias{ne_states}
 \title{Get natural earth world state (admin level 1) polygons}
 \usage{
