+++ data-raw/reponse_to_ropensci_reviews1.txt
@@ -1,41 +1,42 @@
 
 ## responses to @lmullen
 
-### This package should be at least two packages, and probably three. 
+##### This package should be at least two packages, and probably three. 
 
 done. converted to 3 packages rnaturalearth, rnaturalearthdata & rnaturalearthhires. The first 2 air aimed at CRAN and the 3rd the rOpenSci drat repo)
 
-### @sckott, @karthik, @cboettig: To be able to test these changes effectively, @AndySouth will need to use the rOpenSci drat repository. I'd suggest that the package be provisionally accepted and moved into rOpenSci GitHub and drat repositories.
+##### @sckott, @karthik, @cboettig: To be able to test these changes effectively, @AndySouth will need to use the rOpenSci drat repository. I'd suggest that the package be provisionally accepted and moved into rOpenSci GitHub and drat repositories.
+I'm ready for that stage now if you are.
 
-
-### The README.md should be converted to README.Rmd via devtools::use_readme_rmd().
+##### The README.md should be converted to README.Rmd via devtools::use_readme_rmd().
 done
-### The package should use Appveyor's CI for Windows via devtools::use_appveyor().
-done
-
-### Specify a version number for rgdal, httr, and knitr.
-
-### All the tests pass. I'll check coverage more carefully once the package structure has stabilized. In the meantime, all of the tests which download files should have skip_on_cran().
+##### The package should use Appveyor's CI for Windows via devtools::use_appveyor().
 done
 
-### The vignette should not need download anything from Natural Earth. The code block at the end with two downloads should be set to eval = FALSE.
+##### Specify a version number for rgdal, httr, and knitr.
+I don't know what version number is needed for these packages. Feel like I would just be making one up. Where might I look ?
+
+##### All the tests pass. I'll check coverage more carefully once the package structure has stabilized. In the meantime, all of the tests which download files should have skip_on_cran().
 done
 
-### Code style
-
-### In the README and vignettes, in place of require(rnaturalearth) use library(rnaturalearth).
+##### The vignette should not need download anything from Natural Earth. The code block at the end with two downloads should be set to eval = FALSE.
 done
-### Spacing is often inconsistent around = using in function definitions and calls. There should always be a single space on both sides of =. For example, in the README vignette("rnaturalearth", package="rnaturalearth") should be vignette("rnaturalearth", package = "rnaturalearth").
+
+##### Code style
+
+##### In the README and vignettes, in place of require(rnaturalearth) use library(rnaturalearth).
+done
+##### Spacing is often inconsistent around = using in function definitions and calls. There should always be a single space on both sides of =. For example, in the README vignette("rnaturalearth", package="rnaturalearth") should be vignette("rnaturalearth", package = "rnaturalearth").
 I recognise this, but unless it's essential would prefer to keep my sometimes idiosyncratic use of spaces which I use to enhance my readability (and out of habit too).
 
-### Comments at the start of the line should be followed by a space. (I prefer to user sentence capitalization; but that's just a preference.) 
+##### Comments at the start of the line should be followed by a space. (I prefer to user sentence capitalization; but that's just a preference.) 
 done
-### Most important, all code which has been commented out needs to be deleted.
+##### Most important, all code which has been commented out needs to be deleted.
 done (mostly)
-### The same goes for the data download scripts in data-raw. At the moment, they have a lot investigating the data. Now that you've settled on what to do, you can retain that information in comments but trim unnecessary code.
+##### The same goes for the data download scripts in data-raw. At the moment, they have a lot investigating the data. Now that you've settled on what to do, you can retain that information in comments but trim unnecessary code.
 I have kept some commented code as these are the areas where it is most likely I will need to adapt old code in repsonse to potential future data changes, and least likely that users need to access. 
 
[…]
