### 2016-03-21 7ef829b
add seealso to raster::getData
