### 2016-03-21 f70d1a9
fix destdir() bug
