### 2016-03-21 da3dea8
improve raster doc
