+++ R/ne_download.r
@@ -58,16 +58,16 @@
 #' spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
 #' spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )
 #'
-#' # for raster
+#' # for raster, here an example with Manual Shaded Relief (MSR)
 #' # download & load
-#' rst <- ne_download(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' rst <- ne_download(scale = 50, type = "MSR_50M", category = "raster", destdir = getwd())
 #' 
 #' # load after having downloaded
-#' rst <- ne_load(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+#' rst <- ne_load(scale = 50, type = "MSR_50M", category = "raster", destdir = getwd())
 #' 
 #' # plot
 #' library(raster)
-#' plot(rst)
+#' raster::plot(rst)
 #' } # end dontrun
 #' @return A \code{Spatial} object depending on the data (points, lines, polygons or raster), 
 #'    unless load=FALSE in which case it returns the name of the downloaded shapefile (without extension).
