+++ man/ne_download.Rd
@@ -75,16 +75,16 @@ spdf_world2 <-    ne_load( scale = 110, type = 'countries' )
 spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
 spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )
 
-# for raster
+# for raster, here an example with Manual Shaded Relief (MSR)
 # download & load
-rst <- ne_download(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+rst <- ne_download(scale = 50, type = "MSR_50M", category = "raster", destdir = getwd())
 
 # load after having downloaded
-rst <- ne_load(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+rst <- ne_load(scale = 50, type = "MSR_50M", category = "raster", destdir = getwd())
 
 # plot
 library(raster)
-plot(rst)
+raster::plot(rst)
 } # end dontrun
 }
 \seealso{
