### 2016-03-21 9a8fc89
use match.arg for type
