+++ man/get_data.Rd
@@ -4,7 +4,7 @@
 \alias{get_data}
 \title{Get data from within the package}
 \usage{
-get_data(scale = 110, type = "countries")
+get_data(scale = 110, type = c("countries", "map_units", "sovereignty"))
 }
 \arguments{
 \item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10}, \code{'small'}, \code{'medium'}, \code{'large'}}
