### 2016-03-21 cdfde70
make more concise
