+++ R/ne_file_name.r
@@ -40,7 +40,6 @@ ne_file_name <- function(scale = 110,
   
   # add admin_1 to known types
   # this actually just expands 'states' to the name including lakes
-  # todo think about this one
   if (type == 'states')
     type <- 'admin_1_states_provinces_lakes'
 
