+++ R/ne_countries.r
@@ -1,9 +1,9 @@
 #' Get natural earth world country polygons
 #'
-#' returns world country polygons at a specified scale
+#' returns world country polygons at a specified scale, or points of tiny_countries
 #'
 #' @param scale scale of map to return, one of \code{110}, \code{50}, \code{10} or \code{'small'}, \code{'medium'}, \code{'large'}
-#' @param type country type, one of 'countries', 'map_units', 'sovereignty'
+#' @param type country type, one of 'countries', 'map_units', 'sovereignty', 'tiny_countries'
 #' @param continent a character vector of continent names to get countries from.
 #' @param country a character vector of country names. 
 #' @param geounit a character vector of geounit names. 
@@ -15,13 +15,15 @@
 #' spdf_world <- ne_countries()
 #' spdf_africa <- ne_countries(continent='africa')
 #' spdf_france <- ne_countries(country='france')
+#' spdf_tiny_countries <- ne_countries(type='tiny_countries', scale=50)
 #' 
 #' if (require(sp)) {
 #'   plot(spdf_world)
 #'   plot(spdf_africa)
-#'   plot(spdf_france)   
+#'   plot(spdf_france)
+#'   plot(spdf_tiny_countries)   
 #' }
-#' @return A \code{SpatialPolygonsDataFrame} object.
+#' @return \code{SpatialPolygonsDataFrame} or \code{SpatialPointsDataFrame} object.
 #' @export
 #' 
 ne_countries <- function(scale = 110,
