+++ R/get_data.r
@@ -3,7 +3,7 @@
 #' returns world country polygons at a specified scale, used by ne_countries()
 #'
 #' @param scale scale of map to return, one of \code{110}, \code{50}, \code{10}, \code{'small'}, \code{'medium'}, \code{'large'}
-#' @param type country type, one of 'countries', 'map_units', 'sovereignty'
+#' @param type country type, one of 'countries', 'map_units', 'sovereignty', 'tiny_countries'
 # @examples
 # spdf_world <- get_data(scale = 110, type = 'countries')
 #' @return A \code{SpatialPolygonsDataFrame} object.
@@ -11,13 +11,17 @@
 # @export
 #' 
 get_data <- function(scale = 110,
-                     type = c('countries', 'map_units', 'sovereignty') ) {
+                     type = c('countries', 'map_units', 'sovereignty', 'tiny_countries') ) {
 
   # check on permitted scale arg, convert names to numeric
   scale <- check_scale(scale)
   
   # check permitted type arg
   type <- match.arg(type) 
+  
+  # tiny_countries not available at scale 10
+  if ( type=='tiny_countries' && scale==10 )
+    stop('tiny_countries are not available at scale 10, use scale 50 or 110')
   
   # check for the data packages and try to install if not there
   # avoid this check for one example dataset in this package (i.e. countries110)
