+++ man/get_data.Rd
@@ -4,12 +4,13 @@
 \alias{get_data}
 \title{Get data from within the package}
 \usage{
-get_data(scale = 110, type = c("countries", "map_units", "sovereignty"))
+get_data(scale = 110, type = c("countries", "map_units", "sovereignty",
+  "tiny_countries"))
 }
 \arguments{
 \item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10}, \code{'small'}, \code{'medium'}, \code{'large'}}
 
-\item{type}{country type, one of 'countries', 'map_units', 'sovereignty'}
+\item{type}{country type, one of 'countries', 'map_units', 'sovereignty', 'tiny_countries'}
 }
 \value{
 A \code{SpatialPolygonsDataFrame} object.
