### 2016-03-22 #22 @lmullen
@AndySouth Glad to see the next version of this package.

@karthik and @sckott: I can review this package in more detail now (not anticipating any problems) but perhaps it would be best to have the high resolution data in drat so I can check that too.

@AndySouth: The version numbers for rgdal etc should be whatever the current version is, since that is what you are developing against. You can get the current versions from `packageVersion()` or better yet from looking at CRAN.
