### 2016-03-22 #22 @sckott
hmm, I don't know the process for adding repos to our drat, asked over there https://github.com/ropensci/drat/issues/10
