### 2016-03-22 #22 @sckott
@lmullen so we just need `rnaturalearthhires` in drat for now?
