### 2016-03-22 #22 @lmullen
Yes, I think that should work, right @AndySouth? The other two will be on
CRAN, so I can install them as if they were on CRAN. Then test with and
without the high res package.

On Tue, Mar 22, 2016 at 2:38 PM, Scott Chamberlain <notifications@github.com
> wrote:

> @lmullen <https://github.com/lmullen> so we just need rnaturalearthhires
> in drat for now?
>
