### 2016-03-23 #22 @sckott
hmm, i can't get drat to work for `rnaturalearthhires` - @cboettig not sure what's going on, would it help if we move any pkgs into ropensci org account?
