### 2016-03-23 #22 @cboettig
@sckott um, not your fault I think; it loks like CircleCI is failing to build drat, so the package repo on amazon isn't getting updated.  Circle CI seems to be having some issues reading the environmental variables all of a sudden, they have a note about the issue up, hope it is resolved soon.  May need to run drat locally once to avoid timeout on the S3 download logs  too
