### 2016-03-24 #22 @cboettig
@sckott looks like rnaturalearthhires is up on drat now:

http://packages.ropensci.org/src/contrib/PACKAGES shows:

```
Package: rnaturalearthhires
Version: 0.0.0.9000
Depends: R (>= 3.1.1)
Imports: sp (>= 1.0.15)
Suggests: knitr, testthat (>= 0.9.1)
License: CC0
MD5sum: 4b7c1c00c6e4a8ce553f4d0867cf3e2a
NeedsCompilation: no
```
