### 2016-03-24 #22 @sckott
@AndySouth @lmullen `rnaturalearthhires` is working on ropensci drat now. Do we need to add the other  pkgs as well, or one of them?
