### 2016-03-29 #22 @AndySouth
@sckott @lmullen Thanks Scott. Will you want the other packages there if they are accepted to rOpensci ? More broadly, my thought was that I would wait for further comments that Lincoln offered and, when I have time, submit the two smaller packages to Cran.
