### 2016-03-29 #22 @lmullen
@sckott Pretty sure I just need `rnaturalearthhires` on drat if the other
two are intended for CRAN.

@AndySouth: I'll take another look at the packages as soon as I can. (I'm
teaching mapping on Wednesday: maybe I'll even use them in class!)

On Tue, Mar 29, 2016 at 5:29 AM, Andy South <notifications@github.com>
wrote:

> @sckott <https://github.com/sckott> @lmullen <https://github.com/lmullen>
> Thanks Scott. Will you want the other packages there if they are accepted
> to rOpensci ? More broadly, my thought was that I would wait for further
> comments that Lincoln offered and, when I have time, submit the two smaller
> packages to Cran.
>
