### 2016-03-29 #22 @sckott
> Will you want the other packages there if they are accepted to rOpensci ?

by _there_ @AndySouth  do you mean _drat_?  if they are accepted in our suite, they'll be added to our drat
