### 2016-03-29 #22 @AndySouth
@sckott thanks, yes I did mean drat.

On 29 March 2016 at 16:55, Scott Chamberlain <notifications@github.com>
wrote:

> Will you want the other packages there if they are accepted to rOpensci ?
>
> by *there* @AndySouth <https://github.com/AndySouth> do you mean *drat*?
> if they are accepted in our suite, they'll be added to our drat
>
