### 2016-03-30 #22 @AndySouth
@lmullen @Robinlovelace would you like to be added to authors for the package ? (if that's allowed). I think your contributions warrant it and I would like to keep it collaborative. You'd be welcome to make any or no further contributions in future. Up to you either way.
