### 2016-03-30 #22 @lmullen
@AndySouth That's a generous offer.  But I think that for the integrity of the peer review process, it would be better to maintain the distinction between author and reviewer, even though the strength of this form of review is that it is collaborative rather than adversarial. (I don't want to speak for @Robinlovelace, of course, who may disagree.) And in my own case, I haven't contributed any actual code as of yet.
