### 2016-03-30 #22 @AndySouth
@lmullen thanks I can see your point, I guess I wasn't thinking of this as peer review in the usual sense. I'm happy to go with what you think best. I do think of you as having contributed code as I copy/pasted parts of your code that you pointed me towards and modified them only minorly.
