### 2016-03-30 #22 @Robinlovelace
Agree with @lmullen that I've not contributed any code so am not a package author. There's an ongoing debate about how to give credit for the important work of peer review, so perhaps saying somewhere "Reviewed by ..." would be good. Interestingly R provides many types of package contributors, including thesis advisor! (see `?person`).

Delving further I tried the following:

```
> person("Lincoln", "Mullen", role = c("rev"))
[1] "Lincoln Mullen [rev]"
> person("Lincoln", "Mullen", role = c("reviewer"))
[1] "Lincoln Mullen"
Warning message:
In .canonicalize_person_role(role) :
  Invalid role specification: ‘reviewer’.
```
So `rev` works, guess it means 'package reviewer' (although does not say so in the help) and I'd be stoked to mentioned in that way. What are your thoughts on this way of doing in @lmullen  and @sckott (as a potential template for the future). Either way it's been a great experience and illustrates how much peer review of code is beneficial.
