### 2016-05-04 badc4a5
added credit to reviewers although that does prompt a note in CHECK : * checking DESCRIPTION meta-information ... NOTE Authors@R field gives persons with no valid roles:   Lincoln Mullen [rev]   Robin Lovelace [rev]
