### 2016-05-04 #22 @sckott
@AndySouth looking back through comments, was I supposed to do something?  this does work now `install.packages("rnaturalearthhires", repos = c("http://packages.ropensci.org", "http://cran.rstudio.com"))`
