### 2016-05-04 #22 @lmullen
@sckott @AndySouth: Sorry I've been slow. Got hung up with the end of the semester. I'll review the revised package soon.
