### 2016-05-19 #22 @lmullen
@AndySouth: Nice work on this package. I think it is very close to being done and accepted into rOpenSci, then on its way to CRAN. Here are a few minor fixes or questions I found.

## R CMD check

I get the following note, which should be easily corrected:

```
* checking R code for possible problems ... NOTE
ne_download: no visible global function definition for ‘download.file’
ne_download: no visible global function definition for ‘unzip’
  download.file unzip
Consider adding
  importFrom("utils", "download.file", "unzip")
to your NAMESPACE file.
```

## Separation of packages

The separation of packages seems to work fine. I was a little surprised that the package installs the data packages using devtools instead of using the rOpenSci repository. What is the reason for that choice? I'd prefer that the rOpenSci repository was used if possible.

The DESCRIPTION will need this line added:

```
Additional_repositories: http://packages.ropensci.org
```

I'd still like to see package version numbers for the dependencies in the DESCRIPTION.

## R CMD check

In addition, R CMD check fails for me when the rnaturalearthhires package is not available. This is because it can't build the vignette.

The vignette fails to build if rnaturalearthhires is not available. Since that package won't be CRAN, I don't think this is going to pass CRAN checks. You should probably add `eval = FALSE` to the chunks in the vignette that use the high resolution data. It is sufficient to show what the code looks like without actually running it.

I get a warning message when running the code in the examples in `ne_download()`. When I run

```
spdf_world2 <-    ne_load( scale = 110, type = 'countries' )
```

I get this warning message which does not seem relevant to the command run.

```
Warning message:
In if (category == "raster") { :
  the condition has length > 1 and only the first element will be used
```

The reason is that the default argument for `category` is a vector of three choices. This function should probably have the line

```
category <- match.arg(category)
```

## Testing / continuous integration

I noticed that the Travis and Appveyor tests are failing. At least the Travis tests should be passing. It looks like the problem is that the installation of rgdal fails because you don't have the rgdal dev package installed on Ubuntu. I'm not sure why the binary package as specified in the Travis file is not working. But you should be able to install rgdal from source if you install the apt package `libgdal-dev`. As for Appveyor, it looks like the problem has to do with package installation as well.

Not a requirement, but you might consider adding code coverage with [covr](https://cran.rstudio.com/web/packages/covr/).

## Stylistic considerations

I've submitted some very minor clean up to the introductory vigentte in a PR.

I'd suggest changing the title of the introductory vignette to "Introduction to rnaturalearth."

The package is inconsistent in how it handles spaces around ` = ` in function arguments. All user-facing documentation (i.e., examples and vignettes) should be made consistent in having a single space on both sides of the `=`. There are a few places with inconsistent spacing around parentheses as well.

That's about it. Nice work!
