### 2016-05-20 #22 @AndySouth
Many thanks @lmullen for the thorough 2nd review. I'll sort these issues over the next couple of weeks.
