### 2016-06-01 d45f376
add rOpenSci repo
