### 2016-06-01 9537276
add import of utils to correct R CMD check note seen by @lmullen
