+++ DESCRIPTION
@@ -17,7 +17,8 @@ Depends:
     R (>= 3.1.1)
 Imports:
     sp (>= 1.0.15),
-    rgdal (>= 1.1-3)
+    rgdal (>= 1.1-3),
+    utils (>= 3.2.3)
 Suggests:
     knitr (>= 1.12.3),
     testthat (>= 0.9.1),
