### 2016-06-01 9442b0b
to correct warning : condition has length > 1 noted by @lmullen
