### 2016-06-01 58dcf06
not evaluated because rely on rnaturalearthhires data which are on rOpenSci so CRAN check likely to fail
