+++ vignettes/rnaturalearth.Rmd
@@ -54,15 +54,21 @@ sp::plot(ne_countries(country = 'united kingdom', type='countries'))
 # map_units, UK divided into England, Scotland, Wales and Northern Ireland
 sp::plot(ne_countries(country = 'united kingdom', type='map_units'))
 
-
-# countries, large scale
-sp::plot(ne_countries(country = 'united kingdom', scale='large'))
-
-# countries, medium scale
-sp::plot(ne_countries(country = 'united kingdom', scale='medium'))
      
 # countries, small scale
 sp::plot(ne_countries(country = 'united kingdom', scale='small'))   
+
+# countries, medium scale
+sp::plot(ne_countries(country = 'united kingdom', scale='medium'))
+
+```
+
+
+```{r, eval=FALSE, echo=TRUE, message=FALSE}
+# not evaluated because rely on rnaturalearthhires data which are on rOpenSci so CRAN check likely to fail
+
+# countries, large scale
+sp::plot(ne_countries(country = 'united kingdom', scale='large'))
 
 # states country='united kingdom'
 sp::plot(ne_states(country ='united kingdom'))  
@@ -70,7 +76,11 @@ sp::plot(ne_states(country ='united kingdom'))
 sp::plot(ne_states(geounit='england')) 
 
 # states country='france'
-sp::plot(ne_states(country ='france'))  
+sp::plot(ne_states(country ='france'))
+
+```
+
+```{r, eval=TRUE, echo=TRUE, message=FALSE}
 
 # coastline of the world
 # subsetting of coastline is not possible because the Natural Earth data are not attributed in that way
