+++ vignettes/rnaturalearth.Rmd
@@ -1,12 +1,12 @@
 ---
-title: "rnaturalearth vignette"
+title: "Introduction to rnaturalearth."
 author: "Andy South"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 #to produce a pdf
 #output: rmarkdown::pdf_document
 vignette: >
-  %\VignetteIndexEntry{rnaturalearth vignette}
+  %\VignetteIndexEntry{Introduction to rnaturalearth.}
   %\VignetteEngine{knitr::rmarkdown}
   \usepackage[utf8]{inputenc}
 ---
