### 2016-06-01 21251e5
add spaces around =, eval=FALSE for rnaturalearthhires
