+++ vignettes/what-is-a-country.Rmd
@@ -41,57 +41,57 @@ Filtering by `geounit` can give finer control, e.g. to plot Scotland alone, or F
 ```{r, eval=TRUE, echo=TRUE, message=FALSE}
 
 # countries, UK undivided
-sp::plot(ne_countries(country = 'united kingdom', type='countries'))
+sp::plot(ne_countries(country = 'united kingdom', type = 'countries'))
 # map_units, UK divided into England, Scotland, Wales and Northern Ireland
-sp::plot(ne_countries(country = 'united kingdom', type='map_units'))
+sp::plot(ne_countries(country = 'united kingdom', type = 'map_units'))
 # map_units, select by geounit to plot Scotland alone
-sp::plot(ne_countries(geounit = 'scotland', type='map_units'))
+sp::plot(ne_countries(geounit = 'scotland', type = 'map_units'))
 # sovereignty, Falkland Islands included in UK
-sp::plot(ne_countries(country = 'united kingdom', type='sovereignty'), col='red')
-sp::plot(ne_coastline(scale=110), col='lightgrey', lty=3, add=TRUE)
+sp::plot(ne_countries(country = 'united kingdom', type = 'sovereignty'), col = 'red')
+sp::plot(ne_coastline(scale = 110), col = 'lightgrey', lty = 3, add = TRUE)
 
 # France, country includes French Guiana
 sp::plot(ne_countries(country = 'france'))
 # France map_units includes French Guiana too
-sp::plot(ne_countries(country = 'france', type='map_units'))
+sp::plot(ne_countries(country = 'france', type = 'map_units'))
 # France filter map_units by geounit to exclude French Guiana
-sp::plot(ne_countries(geounit = 'france', type='map_units'))
+sp::plot(ne_countries(geounit = 'france', type = 'map_units'))
 # France sovereignty includes South Pacicic islands
-sp::plot(ne_countries(country = 'france', type='sovereignty'), col='red')
-sp::plot(ne_coastline(scale=110), col='lightgrey', lty=3, add=TRUE)
+sp::plot(ne_countries(country = 'france', type = 'sovereignty'), col = 'red')
+sp::plot(ne_coastline(scale = 110), col = 'lightgrey', lty = 3, add = TRUE)
 
 ```
 
 ### Country scales : small, medium and large.
 The different definitions of a country outlined above are available at different scales.
-```{r, eval=TRUE, echo=TRUE, message=FALSE}
+```{r, eval=FALSE, echo=TRUE, message=FALSE}
 
 # countries, large scale
 sp::plot(ne_countries(country = 'united kingdom', scale='large'))
 
 # countries, medium scale
 # temporarily commented out to avoid utf conversion travis error with Curacao
-# sp::plot(ne_countries(country = 'united kingdom', scale='medium'))
+# sp::plot(ne_countries(country = 'united kingdom', scale = 'medium'))
      
 # countries, small scale
-sp::plot(ne_countries(country = 'united kingdom', scale='small'))     
+sp::plot(ne_countries(country = 'united kingdom', scale = 'small'))     
      
      
 ```
 
 
 ### States, admin level 1, select by country or geounit. 
-```{r, eval=TRUE, echo=TRUE, message=FALSE}
+```{r, eval=FALSE, echo=TRUE, message=FALSE}
 
 # states country='united kingdom'
-sp::plot(ne_states(country ='united kingdom'))  
+sp::plot(ne_states(country = 'united kingdom'))  
 # states geounit='england'
-sp::plot(ne_states(geounit='england')) 
+sp::plot(ne_states(geounit = 'england')) 
 
 # states country='france'
-sp::plot(ne_states(country ='france'))  
[…]
