+++ vignettes/rnaturalearth.Rmd
@@ -47,7 +47,7 @@ Pre-downloaded maps can be accessed with :
 ```{r, eval=TRUE, echo=TRUE, message=FALSE}
 
 # world at small scale (low resolution)
-sp::plot(ne_countries(type='countries', scale='small'))
+sp::plot(ne_countries(type = 'countries', scale = 'small'))
 
 # countries, UK undivided
 sp::plot(ne_countries(country = 'united kingdom', type='countries'))
@@ -56,10 +56,10 @@ sp::plot(ne_countries(country = 'united kingdom', type='map_units'))
 
      
 # countries, small scale
-sp::plot(ne_countries(country = 'united kingdom', scale='small'))   
+sp::plot(ne_countries(country = 'united kingdom', scale = 'small'))   
 
 # countries, medium scale
-sp::plot(ne_countries(country = 'united kingdom', scale='medium'))
+sp::plot(ne_countries(country = 'united kingdom', scale = 'medium'))
 
 ```
 
@@ -68,15 +68,15 @@ sp::plot(ne_countries(country = 'united kingdom', scale='medium'))
 # not evaluated because rely on rnaturalearthhires data which are on rOpenSci so CRAN check likely to fail
 
 # countries, large scale
-sp::plot(ne_countries(country = 'united kingdom', scale='large'))
+sp::plot(ne_countries(country = 'united kingdom', scale = 'large'))
 
 # states country='united kingdom'
-sp::plot(ne_states(country ='united kingdom'))  
+sp::plot(ne_states(country = 'united kingdom'))  
 # states geounit='england'
-sp::plot(ne_states(geounit='england')) 
+sp::plot(ne_states(geounit = 'england')) 
 
 # states country='france'
-sp::plot(ne_states(country ='france'))
+sp::plot(ne_states(country = 'france'))
 
 ```
 
@@ -98,12 +98,12 @@ Each [Natural Earth](http://www.naturalearthdata.com/) dataset is characterised 
 ```{r, eval=FALSE, echo=TRUE, message=FALSE}
 
 # lakes
-lakes110 <- ne_download(scale=110, type='lakes', category='physical')
-sp::plot(lakes110, col='blue')
+lakes110 <- ne_download(scale = 110, type = 'lakes', category = 'physical')
+sp::plot(lakes110, col = 'blue')
 
 # rivers
-rivers110 <- ne_download(scale=110, type='rivers_lake_centerlines', category='physical')
-sp::plot(rivers110, col='blue')
+rivers110 <- ne_download(scale = 110, type = 'rivers_lake_centerlines', category = 'physical')
+sp::plot(rivers110, col = 'blue')
 
 
 ```
