### 2016-06-01 5ec8283
make spaces & ' consistent in user-facing stuff
