+++ R/ne_states.r
@@ -12,15 +12,15 @@
 #' @examples
 #' 
 #' # comparing using country and geounit to filter
-#' spdf_france_country <- ne_states(country='france')
-#' spdf_france_geounit <- ne_states(geounit='france')
+#' spdf_france_country <- ne_states(country = 'france')
+#' spdf_france_geounit <- ne_states(geounit = 'france')
 #'  
 #' if (require(sp)) {
 #'   plot(spdf_france_country)
 #'   plot(spdf_france_geounit) 
 #'   
-#'   plot(ne_states(country ='united kingdom'))  
-#'   plot(ne_states(geounit='england'))     
+#'   plot(ne_states(country = 'united kingdom'))  
+#'   plot(ne_states(geounit = 'england'))     
 #' }
 #' @return A \code{SpatialPolygonsDataFrame} object.
 #' @export
