+++ man/ne_load.Rd
@@ -12,7 +12,7 @@ ne_load(scale = 110, type = "countries", category = c("cultural",
 
 \item{type}{type of natural earth file one of 'countries', 'map_units', 'map_subunits', 'sovereignty', 'states'
 OR the portion of any natural earth vector url after the scale and before the . 
-e.g. for "ne_50m_urban_areas.zip" this would be "urban_areas"}
+e.g. for 'ne_50m_urban_areas.zip' this would be 'urban_areas'}
 
 \item{category}{one of natural earth categories : 'cultural', 'physical', 'raster'}
 
@@ -38,10 +38,10 @@ spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )
 
 # for raster
 # download & load
-rst <- ne_download(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+rst <- ne_download(scale = 50, type = 'OB_50M', category = 'raster', destdir = getwd())
 
 # load after having downloaded
-rst <- ne_load(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
+rst <- ne_load(scale = 50, type = 'OB_50M', category = 'raster', destdir = getwd())
 
 # plot
 library(raster)
