+++ man/ne_states.Rd
@@ -25,15 +25,15 @@ returns state polygons (administrative level 1) for specified countries
 \examples{
 
 # comparing using country and geounit to filter
-spdf_france_country <- ne_states(country='france')
-spdf_france_geounit <- ne_states(geounit='france')
+spdf_france_country <- ne_states(country = 'france')
+spdf_france_geounit <- ne_states(geounit = 'france')
  
 if (require(sp)) {
   plot(spdf_france_country)
   plot(spdf_france_geounit) 
   
-  plot(ne_states(country ='united kingdom'))  
-  plot(ne_states(geounit='england'))     
+  plot(ne_states(country = 'united kingdom'))  
+  plot(ne_states(geounit = 'england'))     
 }
 }
 

[…]
