+++ NAMESPACE
@@ -12,3 +12,4 @@ export(ne_load)
 export(ne_states)
 import(rgdal)
 import(sp)
+import(utils)
