### 2016-06-01 b5be58e
start response to 2nd review
