### 2016-06-02 fe96747
trying to fix travis rgdal fail, make same as repijson which currently passes
