### 2016-06-02 0696f7d
move rgdal from import to suggest partly to try to sort travis
