+++ NAMESPACE
@@ -10,6 +10,5 @@ export(ne_download)
 export(ne_file_name)
 export(ne_load)
 export(ne_states)
-import(rgdal)
 import(sp)
 import(utils)
