### 2016-06-02 7ec3b8f
make more like rnoaa
