+++ appveyor.yml
@@ -1,5 +1,9 @@
 # DO NOT CHANGE the "init" and "install" sections below
 
+environment:
+  R_CHECK_ARGS: --no-build-vignettes --no-manual
+  _R_CHECK_FORCE_SUGGESTS_: FALSE
+
 # Download script file from GitHub
 init:
   ps: |
@@ -14,13 +18,13 @@ install:
 
 build_script:
   - travis-tool.sh install_deps
+  - travis-tool.sh r_binary_install proto rgdal
 
 test_script:
   - travis-tool.sh run_tests
 
 on_failure:
-  - 7z a failure.zip *.Rcheck\*
-  - appveyor PushArtifact failure.zip
+  - travis-tool.sh dump_logs
 
 artifacts:
   - path: '*.Rcheck\**\*.log'
