### 2016-06-02 90784a4
add appveyor github installs
