### 2016-06-02 e84c049
try another rgdal travis fix pre-installing gdal-dev stuff
