+++ .travis.yml
@@ -7,8 +7,10 @@ sudo: required
 
 #try to fix rgdal fail
 #2/6/16 not used in repijson which passes
-#before_install:
 #  - sudo apt-get install libv8-dev
+#2/6/16 try something similar to rnoaa without netcdf
+before_install:
+- sudo apt-get install -y gdal-bin libgdal-dev libgdal1-dev libproj-dev libv8-dev
 
 env:
  global:
