### 2016-06-02 1e25d15
another travis rgdal try with fewer installs
previous gave gpg: requesting key E084DAB9 from hkp server keyserver.ubuntu.com
gpgkeys: key E084DAB9 not found on keyserver
