+++ .travis.yml
@@ -10,7 +10,8 @@ sudo: required
 #  - sudo apt-get install libv8-dev
 #2/6/16 try something similar to rnoaa without netcdf
 before_install:
-- sudo apt-get install -y gdal-bin libgdal-dev libgdal1-dev libproj-dev libv8-dev
+#- sudo apt-get install -y gdal-bin libgdal-dev libgdal1-dev libproj-dev libv8-dev
+- sudo apt-get install libgdal-dev libgdal1-dev
 
 env:
  global:
