### 2016-06-02 2a2ef93
travis rgdal add back in -y to try to fix
The following packages have unmet dependencies: libgdal-dev : Depends: libhdf5-serial-dev
