### 2016-06-02 672a056
travis rgdal remove libgdal1-dev to try to fix The following packages have unmet dependencies:  libgdal-dev : Depends: libhdf5-serial-dev
