### 2016-06-02 7dbad4c
travis rgdal back to just binary
