+++ .travis.yml
@@ -9,9 +9,9 @@ sudo: required
 #2/6/16 not used in repijson which passes
 #  - sudo apt-get install libv8-dev
 #2/6/16 try something similar to rnoaa without netcdf
-before_install:
+#before_install:
 #- sudo apt-get install -y gdal-bin libgdal-dev libgdal1-dev libproj-dev libv8-dev
-- sudo apt-get install -y libgdal-dev
+#- sudo apt-get install -y libgdal-dev
 
 env:
  global:
