### 2016-06-02 ab78275
update to 2nd revies response, to paste https://github.com/ropensci/onboarding/issues/22
