### 2016-06-02 #22 @sckott
asking about rgdal, i've been using old `language: c` setup, which works, but means you don't get the faster builds, caching, container based infrastucture, etc. - asking jim hester
