### 2016-06-02 #22 @sckott
jim pointed me to this issue on travis-ci repo https://github.com/travis-ci/travis-ci/issues/5852

So Edzer uses this setup https://github.com/edzer/gstat/blob/master/.travis.yml to install gdal/geos from source

worth trying `dist: trusty` in your travis file e.g. https://github.com/jhollist/quickmapr/blob/master/.travis.yml#L3
