### 2016-06-02 #22 @lmullen
@AndySouth I looked at the Travis logs and your comments in your Travis config file. Two ideas:

First, Travis does install the Ubuntu `r-cran-rgdal` package as it should, but then it also tries to install `rgdal` from source. That seems to me like a bug in Travis.

Second, I couldn't exactly tell which combination of options you used at any given time. I just tried it on my local version of Ubuntu (16.04) and all that I needed to install was `libgdal-dev` to compile `rgdal` from source. It looks like one of your option has conflicting versions of libgdal. Maybe try it with both the binary version `r-cran-rgdal` as you have now plus installing just `libgdal-dev` with `before_install:`?
