### 2016-06-02 #22 @lmullen
@AndySouth I think you might also want to use `install_github:` to install `rnaturalearthdata`.
