### 2016-06-06 2a81e05
travis rgdal try dist: trusty
