### 2016-06-06 ffced66
travis rgdal previous failed with libgdal-dev : Depends: libhdf5-serial-dev
now following quickmapr more closely
