### 2016-06-06 0d0cc5e
remove version of rgdal
