### 2016-06-06 1a9b89d
travis rgdal, try binary now that I've removed version num from DESCRIPTION
