### 2016-06-06 2513904
travis rgdal continuing frustration repeat try of following https://github.com/jhollist/quickmapr/commit/3c81692b7ee56c00e0199a584079bf087752aeeb
