### 2016-06-06 09f58b4
travis rgdal move libproj before libgdal
prev failed with libgdal-dev : Depends: libhdf5-serial-dev
