### 2016-06-14 #22 @AndySouth
Hi @sckott does it seem that Edzers solution didn't work either ?
68f30e781d42ecd5c312d1caf2483e35c7dd74c3

Because it succeeds on appveyor does this suggest it's a specific Travis problem ? Should this stop proceeding to rOpenSci & CRAN ? I'm keen to proceed if possible.
