### 2016-06-14 #22 @sckott
@AndySouth Yeah, let's move on, couldn't sort it out yet, but not crucial as it seems to pass elsewhere just not on travis.

Approved (see new label on issue)
