+++ README.Rmd
@@ -109,4 +109,4 @@ Thanks to [Lincoln Mullen](https://github.com/lmullen) for code structure inspir
     + e.g. least developed countries etc.
 
     
-    
+[![ropensci\_footer](http://ropensci.org/public_images/github_footer.png)](http://ropensci.org)    
