### 2016-06-14 ee43547
start move to ropensci
