+++ README.Rmd
@@ -35,7 +35,7 @@ The [Natural Earth](http://www.naturalearthdata.com/) website structures vector 
 ### Install rnaturalearth
 
 ```
-devtools::install_github("andysouth/rnaturalearth", build_vignettes=TRUE)
+devtools::install_github("ropenscilabs/rnaturalearth", build_vignettes=TRUE)
 
 ```
 
@@ -78,7 +78,7 @@ vignette('what-is-a-country', package='rnaturalearth')
 ```
 
 ## Reproducible download of Natural Earth data into the package
-[Script](https://github.com/AndySouth/rnaturalearthdata/blob/master/data-raw/data_download_script.r) used to get data into the accompanying data packages.
+[Script](https://github.com/ropenscilabs/rnaturalearthdata/blob/master/data-raw/data_download_script.r) used to get data into the accompanying data packages.
 
 ## Acknowledgements
 Thanks to [Lincoln Mullen](https://github.com/lmullen) for code structure inspiration from [USAboundaries](https://github.com/ropensci/USAboundaries), [Hadley Wickham](https://github.com/hadley) for comments and prompting, [Bob Rudis](https://github.com/hrbrmstr) for answers to stackoverflow questions about downloading Natural Earth data into R. The [Natural Earth team](http://www.naturalearthdata.com/about/contributors/) and [Nathan Kelso](https://github.com/nvkelso) for providing such a great resource.
