+++ NAMESPACE
@@ -1,8 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
+export(")")
+export("devtools::install_github(\"ropenscilabs/rnaturalearthdata\")'")
 export(check_rnaturalearthdata)
 export(check_rnaturalearthhires)
-export(install_rnaturalearthdata)
 export(install_rnaturalearthhires)
 export(ne_coastline)
 export(ne_countries)
