### 2016-06-14 #22 @AndySouth
@sckott
1. hurrah ! Thanks.
2. at the transfer on github I get : You don’t have admin rights to ropenscilabs
